refactor(exception handling): Convert DioExceptionHandler to mixin
Converted the DioExceptionHandler class into a mixin and made its methods static. This change simplifies the usage of exception handling in API calls, removing the need for creating an instance of DioExceptionHandler. The callApi method in ClientExceptionHandler has also been updated to be static. Corresponding changes have been made in test cases as well.