refactor(exception_handler): Convert mixins to classes
Converted `DioExceptionHandler` and `ClientExceptionHandler` from mixin to class. Added new method `callApi_` in both classes, which is now the main implementation of API calls. The original static `callApi` method in `DioExceptionHandler` now uses an instance of the class to call `callApi_`. Also added a UserModel mock for testing purposes and created a new test file for ClientExceptionHandler.